fix: crear directorio collectors/memory con los archivos del collector existente - Closes #219
